Moved if statments to home.js
Added ternary operators into pollen and air cards in home.js
