+++ src/components/Home/Home.js
@@ -10,8 +10,6 @@ import { AiOutlineExperiment, AiTwotoneExperiment } from 'react-icons/ai';
 import { IconContext } from "react-icons";
 import SearchBar from "./SearchBar";
 import { useState, React } from 'react'
-import PollenCard from './PollenCard/PollenCard';
-import AirCard from './AirCard/AirCard';
 
 
 
@@ -22,13 +20,40 @@ export default function Home() {
     //data published {pollenData ? pollenData.city_name : 'loading'} with ternary operator for page load
     //
 
+    const airLevel = () => {
+        if (airQualData.data[0].aqi < 50) {
+            return <div>GOOD</div>
+        } else if (airQualData.data[0].aqi < 100) {
+            return <div>MODERATE</div> 
+        } else if (airQualData.data[0].aqi < 150) {
+            return <div>NOT GOOD</div> 
+        } else if (airQualData.data[0].aqi < 200) {
+            return <div>BAD</div>
+        } else if (airQualData.data[0].aqi < 250 ) {
+            return <div>UNHEALTHY</div>
+        } else {
+            return <div>HAZARDOUS</div>
+        }
+    }
+
+    const pollenLevel = () => {
+        if (pollenData.data[0].mold_level == 1) {
+            return <div>LOW</div>
+        } else if (pollenData.data[0].mold_level == 2) {
+            return<div>MEDIUM</div>
+        } else if (pollenData.data[0].mold_level == 3) {
+            return <div>HIGH</div>
+        } else {
+            return <div>VERY HIGH</div>
+        }
+    }
     
         return (
 
             <>
                 <div>
 
-                    <title>George's Portfolio</title>
+                    <title> Allergy Forecast </title>
                     <meta name="description" content="Generated by create next app" />
                     <meta name="viewport" content="width=device-width, initial-scale=1" />
                     <link rel="icon" href="/favicon.ico" />
@@ -39,9 +64,7 @@ export default function Home() {
                         <section>
 
                             <SearchBar setPollenData={setPollenData} setairQualData={setairQualData} />
-                            <PollenCard setPollenData={setPollenData} />
-                            <AirCard setairQualData={setairQualData} />
-
+                      
                             <div className="flex justify-center">
                                 <div
                                     className="block max-w-sm rounded-lg bg-yellow-400 text-center shadow-lg mx-5">
@@ -60,7 +83,7 @@ export default function Home() {
                                             Highest Pollen Type: {pollenData ? pollenData.data[0].predominant_pollen_type : ''}
                                         </p>
                                         <p className="mb-4 text-base text-neutral-600 ">
-                                            Pollen Level: {pollenData ? pollenData.data[0].mold_level : ''}
+                                            Pollen Level: {pollenData ? pollenLevel() : ''}
                                         </p>
                                     </div>
                                     <div
@@ -91,7 +114,7 @@ export default function Home() {
                                             Overall Air Quality: {airQualData ? airQualData.data[0].aqi : ''}
                                         </p>
                                         <p className="mb-4 text-base text-neutral-600 ">
-                                            Air Level: {airQualData ? airQualData.data[0].aqi : ''}
+                                            Air Level: {airQualData ? airLevel() : ''}
                                         </p>
